Generic type settings

+++ src/core/gfx/Canvas.h
@@ -7,6 +7,8 @@
 #include <cmath>
 #include <cstdint>
 
+#include <iostream>
+
 enum class BlendMode {
     Set,
     Add,
@@ -33,7 +35,7 @@ enum class VerticalAlign {
 
 class Canvas {
 public:
-    Canvas(FrameBuffer8bit &frameBuffer, int &brightness) : //Settings &settings) :
+    Canvas(FrameBuffer8bit &frameBuffer, float &brightness) : //Settings &settings) :
         _frameBuffer(frameBuffer),
         _right(frameBuffer.width() - 1),
         _bottom(frameBuffer.height() - 1),
@@ -44,6 +46,7 @@ public:
 
     uint8_t color() const { return _color; }
     void setColor(uint8_t color) { _color = color * _brightness; }
+//    void setColor(uint8_t color) { std::cout<<_brightness<<std::endl; _color = color * _brightness; }
 
     BlendMode blendMode() const { return _blendMode; }
     void setBlendMode(BlendMode blendMode) { _blendMode = blendMode; }
@@ -267,5 +270,5 @@ private:
     BlendMode _blendMode = BlendMode::Set;
     Font _font = Font::Default;
 //    Settings &_settings;
-    int &_brightness;
+    float &_brightness;
 };
